Migrate Quest Bot sample to use TTK v5
