+++ .gitignore
@@ -17,7 +17,6 @@ pids
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
-/**/.vscode
 /**/lib
 lib
