--- js/samples/04.q.questBot/.env
@@ -1,3 +1,0 @@
-MicrosoftAppId=
-MicrosoftAppPassword=
-OpenAIKey=
