--- js/samples/04.q.questBot/.eslintignore
@@ -1,8 +1,0 @@
-bin
-build
-demo-packages
-dist
-manifest
-node_modules
-package-lock.json
-docs/assets/main.js
