+++ js/samples/04.q.questBot/.gitignore
@@ -1,2 +1,16 @@
+# TeamsFx files
+env/.env.*.user
+env/.env.local
+.localConfigs
+appPackage/build
+
+# dependencies
+node_modules/
+
+# misc
 .env
-*.zip
+.deployment
+.DS_Store
+
+# build
+lib/
