--- js/samples/04.q.questBot/.prettierrc
@@ -1,11 +1,0 @@
-{
-    "arrowParens": "always",
-    "endOfLine": "auto",
-    "printWidth": 120,
-    "semi": true,
-    "singleAttributePerLine": false,
-    "singleQuote": true,
-    "tabWidth": 4,
-    "trailingComma": "none",
-    "useTabs": false
-}
