+++ js/samples/04.q.questBot/appPackage/manifest.json
@@ -2,7 +2,7 @@
     "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.16/MicrosoftTeams.schema.json",
     "version": "1.0.0",
     "manifestVersion": "1.16",
-    "id": "<<MICROSOFT_APP_OR_BOT_ID>>",
+    "id": "${{TEAMS_APP_ID}}",
     "packageName": "com.package.name",
     "name": {
         "short": "DM",
@@ -24,25 +24,15 @@
         "color": "color.png"
     },
     "accentColor": "#FFFFFF",
-    "staticTabs": [
-        {
-            "entityId": "conversations",
-            "scopes": ["personal"]
-        },
-        {
-            "entityId": "about",
-            "scopes": ["personal"]
-        }
-    ],
     "bots": [
         {
-            "botId": "<<YOUR_BOT_ID>>",
-            "scopes": ["personal", "team", "groupChat"],
-            "isNotificationOnly": false,
-            "supportsCalling": false,
-            "supportsVideo": false,
-            "supportsFiles": false
+            "botId": "${{BOT_ID}}",
+            "scopes": [
+                "personal"
+            ]
         }
     ],
-    "validDomains": []
+    "validDomains": [
+        "${{BOT_DOMAIN}}"
+    ]
 }
