+++ js/samples/04.q.questBot/src/actions/index.ts
@@ -1,5 +1,5 @@
 import { Application, OpenAIPlanner } from '@microsoft/botbuilder-m365';
-import { ApplicationTurnState } from '../bot';
+import { ApplicationTurnState } from '../teamsBot';
 import { inventoryAction } from './inventoryAction';
 import { locationAction } from './locationAction';
 import { mapAction } from './mapAction';
