+++ js/samples/04.q.questBot/src/config.ts
@@ -1,0 +1,7 @@
+const config = {
+  botId: process.env.BOT_ID,
+  botPassword: process.env.BOT_PASSWORD,
+  openAIAPIKey: process.env.OPENAI_API_KEY
+};
+
+export default config;
