+++ js/samples/04.q.questBot/src/index.ts
@@ -1,53 +1,52 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
 // Import required packages
-import { config } from 'dotenv';
-import * as path from 'path';
-import * as restify from 'restify';
+import * as restify from "restify";
 
 // Import required bot services.
 // See https://aka.ms/bot-services to learn more about the different parts of a bot.
 import {
-    CloudAdapter,
-    ConfigurationBotFrameworkAuthentication,
-    ConfigurationBotFrameworkAuthenticationOptions,
-    TurnContext
-} from 'botbuilder';
+  CloudAdapter,
+  ConfigurationServiceClientCredentialFactory,
+  ConfigurationBotFrameworkAuthentication,
+  TurnContext,
+} from "botbuilder";
 
-// Read botFilePath and botFileSecret from .env file.
-const ENV_FILE = path.join(__dirname, '..', '.env');
-config({ path: ENV_FILE });
-
-const botFrameworkAuthentication = new ConfigurationBotFrameworkAuthentication(
-    process.env as ConfigurationBotFrameworkAuthenticationOptions
-);
+// This bot's main dialog.
+import * as teamsBot from "./teamsBot";
+import config from "./config";
 
 // Create adapter.
-// See https://aka.ms/about-bot-adapter to learn more about how bots work.
-const adapter = new CloudAdapter(botFrameworkAuthentication);
+// See https://aka.ms/about-bot-adapter to learn more about adapters.
+const credentialsFactory = new ConfigurationServiceClientCredentialFactory({
+  MicrosoftAppId: config.botId,
+  MicrosoftAppPassword: config.botPassword,
+  MicrosoftAppType: "MultiTenant",
+});
 
-// Create storage to use
-//const storage = new MemoryStorage();
+const botFrameworkAuthentication = new ConfigurationBotFrameworkAuthentication(
+  {},
+  credentialsFactory
+);
+
+const adapter = new CloudAdapter(botFrameworkAuthentication);
 
 // Catch-all for errors.
 const onTurnErrorHandler = async (context: TurnContext, error: Error) => {
-    // This check writes out errors to console log .vs. app insights.
-    // NOTE: In production environment, you should consider logging this to Azure
-    //       application insights.
-    console.error(`\n [onTurnError] unhandled error: ${error.toString()}`);
+  // This check writes out errors to console log .vs. app insights.
+  // NOTE: In production environment, you should consider logging this to Azure
+  //       application insights.
+  console.error(`\n [onTurnError] unhandled error: ${error}`);
 
-    // Send a trace activity, which will be displayed in Bot Framework Emulator
-    await context.sendTraceActivity(
-        'OnTurnError Trace',
-        `${error.toString()}`,
-        'https://www.botframework.com/schemas/error',
-        'TurnError'
-    );
+  // Send a trace activity, which will be displayed in Bot Framework Emulator
+  await context.sendTraceActivity(
+    "OnTurnError Trace",
+    `${error}`,
+    "https://www.botframework.com/schemas/error",
+    "TurnError"
+  );
 
-    // Send a message to the user
-    await context.sendActivity('The bot encountered an error or bug.');
-    await context.sendActivity('To continue to run this bot, please fix the bot source code.');
+  // Send a message to the user
+  await context.sendActivity(`The bot encountered unhandled error:\n ${error.message}`);
+  await context.sendActivity("To continue to run this bot, please fix the bot source code.");
 };
 
 // Set the onTurnError for the singleton CloudAdapter.
@@ -56,19 +55,13 @@ adapter.onTurnError = onTurnErrorHandler;
 // Create HTTP server.
 const server = restify.createServer();
 server.use(restify.plugins.bodyParser());
-
 server.listen(process.env.port || process.env.PORT || 3978, () => {
-    console.log(`\n${server.name} listening to ${server.url}`);
-    console.log('\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator');
-    console.log('\nTo test your bot in Teams, sideload the app manifest.json within Teams Apps.');
+  console.log(`\nBot Started, ${server.name} listening to ${server.url}`);
 });
 
-// Listen for incoming server requests and forward to bot.
-import * as bot from './bot';
-server.post('/api/messages', async (req, res) => {
-    // Route received a request to adapter for processing
-    await adapter.process(req, res as any, async (context) => {
-        // Dispatch to application for routing
-        await bot.run(context);
-    });
+// Listen for incoming requests.
+server.post("/api/messages", async (req, res) => {
+  await adapter.process(req, res, async (context) => {
+    await teamsBot.run(context);
+  });
 });
