+++ js/samples/04.q.questBot/src/items.ts
@@ -1,4 +1,4 @@
-import { parseNumber } from './bot';
+import { parseNumber } from './teamsBot';
 import { IItemList } from './interfaces';
 
 /**
