+++ js/samples/04.q.questBot/tsconfig.json
@@ -1,14 +1,14 @@
 {
-    "compilerOptions": {
-        "declaration": true,
-        "forceConsistentCasingInFileNames": true,
-        "incremental": true,
-        "module": "commonjs",
-        "outDir": "./lib",
-        "rootDir": "./src",
-        "sourceMap": true,
-        "strict": true,
-        "target": "es2016",
-        "tsBuildInfoFile": "./lib/.tsbuildinfo"
-    }
+  "compilerOptions": {
+    "declaration": true,
+    "target": "es2017",
+    "module": "commonjs",
+    "outDir": "./lib",
+    "rootDir": "./src",
+    "sourceMap": true,
+    "incremental": true,
+    "tsBuildInfoFile": "./lib/.tsbuildinfo",
+    "resolveJsonModule": true,
+    "esModuleInterop": true,
+  }
 }
